+++ src/pages/CouponTab/CouponTab.jsx
@@ -5,15 +5,21 @@
  * @modify date 2024-02-28 02:18:29
  * @desc 이벤트 탭 나의 QR, coupon
  */
+/**
+ * @author 엄상은
+ * @email sangeun.e.9@gmail.com
+ * @create date 2024-03-04 11:16:31
+ * @modify date 2024-03-04 11:16:31
+ * @desc QR코드 API 연결 및 소켓 연결
+ */
 import React, { useState } from 'react';
 
 //component
 import Header from 'components/Header/Header';
 import CouponCard from 'components/CouponCard/CouponCard';
+import MyQrLoader from 'components/QRScanner/MyQrLoader';
 //hook
 import useGetUserCouponList from 'hooks/useGetUserCouponList';
-//assets
-import myQR from 'assets/images/myQR.png';
 // import bottleIcon from 'assets/icons/bottle_icon.png';
 
 function CouponTab() {
@@ -26,43 +32,6 @@ function CouponTab() {
   const handleTabClick = (activeTab) => {
     setActiveTab(activeTab);
   };
-  //   {
-  //     imgSrc: bottleIcon,
-  //     title: '엔젤리너스 30% 할인 쿠폰',
-  //     detail: '커피 음료 구매시 30% 할인',
-  //     useByDate: '사용 기한 : 2024년 3월 14일까지',
-  //   },
-  //   {
-  //     imgSrc: bottleIcon,
-  //     title: '스타벅스 20% 할인 쿠폰',
-  //     detail: '모든 음료에 사용 가능',
-  //     useByDate: '사용 기한 : 2024년 3월 20일까지',
-  //   },
-  //   {
-  //     imgSrc: bottleIcon,
-  //     title: '엔젤리너스 30% 할인 쿠폰',
-  //     detail: '커피 음료 구매시 30% 할인',
-  //     useByDate: '사용 기한 : 2024년 3월 14일까지',
-  //   },
-  //   {
-  //     imgSrc: bottleIcon,
-  //     title: '스타벅스 20% 할인 쿠폰',
-  //     detail: '모든 음료에 사용 가능',
-  //     useByDate: '사용 기한 : 2024년 3월 20일까지',
-  //   },
-  //   {
-  //     imgSrc: bottleIcon,
-  //     title: '엔젤리너스 30% 할인 쿠폰',
-  //     detail: '커피 음료 구매시 30% 할인',
-  //     useByDate: '사용 기한 : 2024년 3월 14일까지',
-  //   },
-  //   {
-  //     imgSrc: bottleIcon,
-  //     title: '스타벅스 20% 할인 쿠폰',
-  //     detail: '모든 음료에 사용 가능',
-  //     useByDate: '사용 기한 : 2024년 3월 20일까지',
-  //   },
-  // ];
 
   return (
     <div className="flex flex-col min-h-screen">
@@ -84,10 +53,7 @@ function CouponTab() {
       <div className="flex flex-grow justify-center ">
         {activeTab === 'QR' ? (
           <div className="flex items-center">
-            <div className="text-center">
-              <img src={myQR} className="h-[35vh] mx-auto" />
-              <p className="font-bold mt-4 mb-32">매장 직원에게 QR코드를 보여주세요.</p>
-            </div>
+            <MyQrLoader />
           </div>
         ) : (
           <div>
